restructured codebase
