+++ .gitignore
@@ -1,7 +1,4 @@
 node_modules
-
 npm-debug.log*
-
-dist
-
 package-lock.json
+.env
