+++ package.json
@@ -20,6 +20,7 @@
   "dependencies": {
     "bcryptjs": "^2.4.3",
     "body-parser": "^1.19.0",
+    "dotenv": "^8.1.0",
     "express": "^4.17.1",
     "google-maps-react": "^2.0.2",
     "jsonwebtoken": "^8.5.1",
