+++ server/controllers/databaseController.js
@@ -1,15 +1,14 @@
-const { Pool } = require('pg');
-const pool = new Pool({
-  connectionString: 'postgres://hiiudcpk:Ozq8Ezo4e1c0HZGONL6i2M5_PmHDKUMV@salt.db.elephantsql.com:5432/hiiudcpk'
-});
+const db = require('../database/database.js');
+
+const databaseController = { };
 
 const SALT_WORK_FACTOR = 10;
 const bcrypt = require('bcryptjs');
 
 // query fetching all comments for specific trails
-const getComment = (req, res, next) => {
+databaseController.getComment = (req, res, next) => {
   const { id } = req.headers;
-  pool.query('SELECT * FROM comments where id = $1', [id], (error, results) => {
+  db.query('SELECT * FROM comments where id = $1', [id], (error, results) => {
     if (error) throw error;
     res.locals.comments = results.rows;
     return next();
@@ -17,13 +16,13 @@ const getComment = (req, res, next) => {
 };
 
 //query posting new comment to DB and then fetching all comments including the one just posted
-const postComment = (req, res, next) => {
+databaseController.postComment = (req, res, next) => {
   const { author, comment, id } = req.body;
   
   if(author && comment && id) {
-    pool.query('INSERT INTO comments (author, comment, id) VALUES ($1, $2, $3)', [author, comment, id], (error, results) => {
+    db.query('INSERT INTO comments (author, comment, id) VALUES ($1, $2, $3)', [author, comment, id], (error, results) => {
     if (error) throw error;
-    pool.query('SELECT * FROM comments where id = $1', [id], (error, results) => {
+    db.query('SELECT * FROM comments where id = $1', [id], (error, results) => {
       if (error) throw error;
       res.locals.comments = results.rows;
       return next();
@@ -33,14 +32,14 @@ const postComment = (req, res, next) => {
 };
 
 //add user and bcrypt password to database
-const createUser = (req, res, next) => {
+databaseController.createUser = (req, res, next) => {
   const { username, password } = req.body;
   if (username && password) {
-    pool.query('SELECT * from users WHERE username = $1', [username], (err, results) => {
+    db.query('SELECT * from users WHERE username = $1', [username], (err, results) => {
       if(results.rows.length === 0) {
         bcrypt.hash(password, SALT_WORK_FACTOR, (err, hash) => {
           if (err) throw err;
-          pool.query('INSERT INTO users (username, password) VALUES ($1, $2) returning *', [username, hash], (error, results) => {
+          db.query('INSERT INTO users (username, password) VALUES ($1, $2) returning *', [username, hash], (error, results) => {
             if (error) throw error;
             res.locals.verified = true;
             return next();
@@ -55,10 +54,10 @@ const createUser = (req, res, next) => {
 };
 
 // query username and password and see if matches are in the database
-const verifyUser = (req, res, next) => {
+databaseController.verifyUser = (req, res, next) => {
   const { username, password } = req.body;
 
-  pool.query('SELECT password FROM users where username = $1', [username], (error, results) => {
+  db.query('SELECT password FROM users where username = $1', [username], (error, results) => {
     if (error) throw error;
     if(results.rows.length === 1) {
       bcrypt.compare(password, results.rows[0].password, (err, isMatch) => {
@@ -77,10 +76,3 @@ const verifyUser = (req, res, next) => {
     };
   });
 };
-
-module.exports = {
-  getComment,
-  verifyUser,
-  createUser,
-  postComment
-}
